+++ docs-site/scripts/generate-playgrounds.R
@@ -36,6 +36,7 @@ playground_heights <- list(
   checkbox = 720L,
   code = 720L,
   empty = 720L,
+  `file-input` = 720L,
   field = 720L,
   icon = 720L,
   input = 720L,